hawk.servlet: move /thrift/hawk into /thrift/hawk/tuple, add /thrift/hawk/{json,binary,compact}
While the TTupleProtocol is the most efficient in terms of space, it is only
available for Java. For almost all the other languages, it is TCompactProtocol,
but it is not available for JavaScript. thrift.js only supports TJSONProtocol
by default, but there is an experimental extension that enables support for
TBinaryProtocol as well.