Implemented most of yesterday's feedback on remote notifications
The EMF, CLI, GUI and URL-based clients can all use one of the four Thrift
protocols supported by Hawk: binary, compact, tuple and JSON. Each of these
protocols is available in a different endpoint:

- /hawk/thrift/binary
- /hawk/thrift/compact
- /hawk/thrift/tuple
- /hawk/thrift/json

The endpoint used will also determine the encoding of the events sent to
the queue mentioned in the reply to #watchModelChanges.

Default and durable queues are now created by the server (since STOMP does
not include methods for remotely creating queues), but temporary queues will
still need to be created by the client. The MONDO server cannot create them,
as they'd be removed automatically as usual when it disconnected from Artemis.

The CLI client now closes subscriptions and connections as well upon exit.