Start integrating Apache Shiro for authentication in Hawk
Apache Shiro seems to be very lightweight and configurable (only requires two
.jar files): this commit adds tha basic integration to make it filter all requests
to the Hawk servlets.

The only problem is that HttpService does not allow for easily integrating the
context listener and init parameters that the regular ShiroFilter requires, so we
are using the deprecated IniShiroFilter that does not require them.

It seems that HttpService creates a separate servlet context for each plugin as
well. This makess sense from an OSGi standpoint, but it does mean that we need
to register a Shiro filter for each plugin that contributes servlets, with its own
shiro.ini file. Authentication can be customized by end users by editing the shiro.ini
files that are placed in the OSGi storage folders on plugin startup (if they are missing).

This is only the first commit that adds Shiro to the mix and fits in with all the
OSGi details, but auth is still disabled by default until we complete two more tasks:

- use the Users Thrift API to manipulate a custom Shiro realm for
  authentication.

- revise the service operation implementations to make them use the
  SecurityUtils.getSubject() call and remove the RequestAware* and
  base HawkThriftServlet classes.