Delay producing Artemis messages until we have real IDs in the graph nodes
During a transaction, new documents in OrientDB only have temporary IDs,
which shouldn't be included in any notification messages. For that reason,
this commit modifies the Artemis message producer to delay the generation
of the messages until the transaction is committed, using message factories.

This will increase memory usage if we're using a subscription over a very
large batch update, but over transactional updates it shouldn't matter,
since OrientDB will be holding on to those documents anyway.