users.servlet can now protect /thrift/** (still disabled by default)
The server.users.servlet plugin provides a filter for all /thrift/**
URLs that can secure them through Apache Shiro. The default config uses
our custom realm, which is backed by a transactional embedded MapDB
database and managed through the Users API.

Usernames and passwords can be now provided through all the command-line
tools. If a Console is available, the tools will be able to read passwords
silently. Otherwise, users will need to specify them directly through the
command (with the risk of shoulder surfing). This is mostly a problem when
using the Eclipse Console view.

Security is however still disabled by default on the default shiro.ini file
because it needs more work before being fully ready for adoption:

- .hawkmodel files need to be able to supply usernames/passwords.
  Ideally, passwords should be optional: if left blank, we'll just
  ask the user and then rely on Eclipse's secure storage for that.
  The problem is what to do if the .hawkmodel is being read from
  a headless environment (e.g. a CloudATL Hadoop job). We'll probably
  need to accept specifying a cleartext password as a fallback in
  that case.

- The ThriftRemote* classes also need to be updated.

- Artemis needs to be integrated into the user management system as
  well. Artemis servers seem to accept a security manager interface
  into which we could plug this account system, but it remains to be
  tested.

Additionally, we still need to test the availability of SecurityUtil
in other .servlet plugins (to check the current user).